Use functors to initialize BVH in single leaf case
The previous version was allocating mirrors on the host, filling, and
copying back to the device. But it was not even quite right because the
bounding boxes live on the device and I had forgotten to copy to the
host. Regardless, since everything is already on the device, it is
easier to execute there rather than copying back and forth.

Also, do not make temp functor object, just pass constructor as argument
to the parallel for.