Disable errors on warnings until figure it out
Right now, the results on any CI build are unpredictable. It may produce
no errors, or it may fail with hundreds of unrelated ones coming from
the system libraries.

I have not had time to investigate it. However, it holds up the process
of merging other PRs. Thus, I'm disabling it for now. It will still
report the warnings, just not fail on them.