Improve PredicateWithAttachment and attach()
Co-authored-by: Damien L-G <dalg24@gmail.com>

The former version did not allow to attach some data, for example, an
int, failing with

  no match for 'operator=' (operand types are
    'ArborX::PredicateWithAttachment<ArborX::Intersects<ArborX::Point>, int>' and
    'ArborX::PredicateWithAttachment<ArborX::Intersects<ArborX::Point>, int&>')

After investigation, we updated implementation with an extra
constructor, and used decay and argument forwarding. These changes
follow clang's implementation of std::pair and std::make_pair. While we
could have added constexpr to the PredicateWithAttachment, we decided
not to, as we do not envision a situation where we do queries at compile
time. Another thing that we skipped was dealing with std::reference.