Properly update hash in version file when building from subdirs
Before, if one ran make from a subdirectory, the record_hash command was
not run, and thus hash was not updated. Thus, if you stayed in the same
subdirectory, but switched branches/hashes, the executable kept having
the same hash (unless checkout also touched CMakeLists.txt).

This commit fixes it by making ArborX target depend on
always-out-of-date record_hash target.