Fixed a bug in DBSCAN leading to bridged clusters
Specifically, we do propagate(i,j) when i is a boundary point, and j is
a core point. But what it really does is that it sets the larger
representative to the smaller one. So what could happen is that the
representative of the core point could be set to the representative of
the boundary point. And that could lead to issue. If boundary point is
then connected to a different cluster, and updates its representative to
that one, this would effectively bridge two clusters, which is a
mistake.