Check that MPI is not finalized before freeing the communicator
We introduced duplicating of user-provided communicators to cleanly
separate the library's communication from that of the user. We store the
duplicated communicator inside ArborX::DistributedSearchTree. However,
what we did not realize was that it can lead to the problem in the
following situation:

MPI_Initialize(argc, argv);
ArborX::DistributedSearchTree<DeviceType> tree(comm, primitives);
tree.query(...);
MPI_Finalize();

The issue is that the destructor for the distributed tree, which
contains MPI_comm_free(&_comm); will be called after MPI_Finalize().

We now check that the MPI is not finalized before calling the function.