Fence ArborX calls in the benchmark
It may happen that the last kernel in ArborX is a Kokkos::parallel_for.
As ArborX does have a guaranteed fence() at the end, what could happen
is that that kernel is launched, and ArborX returns. Thus, when the
ending timer is called, the kernel has not been completed yet. Thus, the
measured time would be wrong in this case.