Reduce default value of cluster_min_size in DBSCAN to 1
This is useful for reproducibility in subtle sitations when a core point
may be stripped of all its border points, resulting in a cluster
consisting of a single core point. As this is triggered by different
algorithms and runs, it makes it challenging to compare the results
based on the number of clusters and the number of points in them.