Make data file sampling portable
Standard does not guarantee performance portability of <random>. While
by some luck random engines are portable (except for the
default_random_engine), uniform_int_distribution is not.

The problem is then that when we sample a dataset on different
platforms/compilers, we get a different data. It is problematic for
reproduciblity.

While it is possible to use something like boost random, I opted instead
to use Lehmer random generator in place, as it is fast, easty to
implement, and does what we need.