Companion to 0fdfa87
In 0fdfa87, we addressed the temporaries issues when passing around
PairIndexObject. What we didn't realize at the time, that similar issues
could have been present in the geometries part too.

This was shown to break behavior of the FDBSCAN-DenseBox when called
with primitives assembled on the fly using AccessTraits. In that case,
the scene bounding box was completely wrong, resulting in random
asserts.