Better way for a user to indicate the intention
The problem with tags is that their behavior is not visible for users.
If a previous version of ArborX ignores tags, it may just silently do
something else.

Instead, we want to fail compilation if the user indicates an intention
we do not support. This is one way to do it: user has to wrap their
callback into a class.